requirements.txt
charset-normalizer==3.4.4
idna==3.11
packaging==26.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-dotenv==1.2.1
qasync==0.28.0
QDarkStyle==3.2.3
QtPy==2.4.3
requests==2.32.5
shiboken6==6.10.2
urllib3==2.6.3